requirements.txt
# Core FastAPI dependencies
fastapi==0.104.1
uvicorn==0.35.0
pydantic==2.5.0
pydantic_core==2.14.1
python-multipart==0.0.20

# ML/AI dependencies
tensorflow==2.16.1
numpy==1.26.4
opencv-python==4.8.1.78
pillow==10.0.1

# Utility dependencies
annotated-types==0.7.0
anyio==3.7.1
click==8.2.1
h11==0.16.0
idna==3.10
sniffio==1.3.1
starlette==0.27.0
typing-inspection==0.4.1
typing_extensions==4.14.1

# DVC for model management
dvc[s3]==3.50.2

# AWS dependencies for S3 storage
boto3==1.39.11
botocore==1.39.11
